rename broadcast receivers' package
