+++ app/src/main/AndroidManifest.xml
@@ -36,17 +36,17 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        
+
         <receiver
-            android:name=".services.BootCompletedReceiver"
+            android:name=".broadcastReceivers.BootCompletedReceiver"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.BOOT_COMPLETED" />
             </intent-filter>
         </receiver>
-        <receiver android:name=".services.AlarmReceiver" />
+        <receiver android:name=".broadcastReceivers.AlarmReceiver" />
         <receiver
-            android:name=".services.AlarmPermissionReceiver"
+            android:name=".broadcastReceivers.AlarmPermissionReceiver"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.app.action.SCHEDULE_EXACT_ALARM_PERMISSION_STATE_CHANGED" />
